+++ domain/model/wallet.go
@@ -8,7 +8,7 @@ import (
 
 type Wallet struct {
 	ID        uuid.UUID `json:"id" msgpack:"id" validate:"required"`
-	Charge    uint64    `json:"charge" msgpack:"charge" validate:"required"`
+	Charge    float64    `json:"charge" msgpack:"charge" validate:"required"`
 	CreatedAt time.Time `json:"createdAt" msgpack:"createdAt"`
 	UpdatedAt time.Time `json:"updatedAt" msgpack:"updatedAt"`
 	User      *User     `json:"user" msgpack:"user" validate:"required"`
